requirements.txt
python-chess>=1.9.4
pandas>=1.5.0
numpy>=1.23.0
tqdm>=4.64.0
scikit-learn>=1.2.0